% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{betas}
\alias{betas}
\title{Calcula os coeficientes (betas) da regressão linear}
\usage{
betas(dados, x, y)
}
\arguments{
\item{dados}{Um data frame ou matriz contendo as variáveis independentes e dependentes.}

\item{x}{Uma matriz das variáveis independentes. Deve ser numérica, sem a coluna do intercepto.}

\item{y}{Um vetor numérico com os valores observados da variável dependente.}
}
\value{
Um vetor numérico com os coeficientes estimados (betas).
}
\description{
Esta função calcula os coeficientes estimados (\(\hat{\beta}\)) da regressão linear
utilizando a fórmula matricial \deqn{\hat{\beta} = (X^T X)^{-1} X^T y}.
}
\examples{
X <- matrix(c(1, 2, 3, 4, 5), ncol = 1)
y <- c(2, 4, 6, 8, 10)
betas_est <- betas(dados = NULL, X, y)
}
