% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{novos_preditos}
\alias{novos_preditos}
\title{Faz a predição de novos valores com base no modelo ajustado}
\usage{
novos_preditos(novox, betas)
}
\arguments{
\item{novox}{Uma matriz ou data frame numérico com os novos valores para predição.
O número de colunas deve ser o mesmo da matriz X original (sem a coluna de intercepto).}

\item{betas}{Uma matriz com os valores dos betas da regressão. Recomenda-se o uso da função 'betas'}
}
\value{
Um vetor numérico com os valores preditos para os novos dados.
}
\description{
Esta função realiza predições para novos dados com base em um modelo de regressão ajustado manualmente.
}
\examples{
newx <- matrix(c(2, 5, 7), ncol = 1)
beta_nx <- matrix(c(0.5, 2), nrow = 2, ncol = 1)
novas_predicoes <- novos_preditos(newx, beta_nx)
}
