% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{preditos}
\alias{preditos}
\title{Calcula os valores preditos do modelo de regressão linear}
\usage{
preditos(x, betas)
}
\arguments{
\item{x}{Uma matriz das variáveis independentes. Deve ser numérica, sem a coluna do intercepto.}

\item{betas}{Uma matriz com os valores dos betas da regressão. Recomenda-se o uso da função 'betas'}
}
\value{
Um vetor numérico com os valores preditos.
}
\description{
Esta função calcula os valores preditos com base nas variáveis independentes
e nos coeficientes estimados pelo modelo de regressão.
}
\examples{
X <- matrix(c(1, 2, 3, 4, 5), ncol = 1)
beta_p <- matrix(c(0.5, 2), nrow = 2, ncol = 1)
y_pred <- preditos(X, beta_p)
}
