% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{residuos}
\alias{residuos}
\title{Calcula os resíduos do modelo de regressão linear}
\usage{
residuos(y, preditos)
}
\arguments{
\item{y}{Um vetor numérico com os valores observados da variável dependente.}

\item{preditos}{Um vetor numérico com os valores preditos da variável dependente. Recomenda-se o uso da função 'preditos'}
}
\value{
Um vetor numérico com os resíduos.
}
\description{
Esta função calcula os resíduos (diferença entre os valores observados e os preditos)
para um modelo de regressão linear.
}
\examples{
y <- c(2, 4, 6, 8, 10)
y_pred <- matrix(c(2.5, 4.6, 6.2, 8.9, 11), nrow = 5, ncol = 1)
res <- residuos(y, y_pred)
}
